Can create random data, can switch between ethernet and random data, and got rid of barometric pressure display
The random data just changes the low level temperature readings by a
random slope until it hits the ranges. The random data uses a timer for
when it will send out the information.

There is no label to display the barometric pressure. And the use can
only select to graph the low level temperature. The charts have defined
sizes and when there are two many points will shift the data over one.

Also added a button to the UI to switch between the random data and
using UDP packets.